split up function for clarity and reusability
