+++ website/src/pages/seq/[accessionVersion].fa/index.ts
@@ -3,7 +3,6 @@ import type { APIRoute } from 'astro';
 import { getReferenceGenomes } from '../../../config.ts';
 import { routes } from '../../../routes/routes.ts';
 import { LapisClient } from '../../../services/lapisClient.ts';
-import { fastaEntryToString, parseFasta } from '../../../utils/parseFasta.ts';
 import { createDownloadAPIRoute } from '../sequenceDownload.ts';
 
 export const GET: APIRoute = createDownloadAPIRoute(
@@ -17,22 +16,7 @@ export const GET: APIRoute = createDownloadAPIRoute(
         const isMultiSegmented = segmentNames.length > 1;
 
         return !isMultiSegmented
-            ? lapisClient.getUnalignedSequences(accessionVersion)
-            : (await lapisClient.getUnalignedSequencesMultiSegment(accessionVersion, segmentNames)).map(
-                  (segmentFastas) =>
-                      segmentFastas
-                          .map((fasta, i) => {
-                              const parsed = parseFasta(fasta);
-                              if (parsed.length === 0) {
-                                  return '';
-                              }
-                              const withSegmentSuffix = {
-                                  name: `${parsed[0].name}_${segmentNames[i]}`,
-                                  sequence: parsed[0].sequence,
-                              };
-                              return fastaEntryToString([withSegmentSuffix]);
-                          })
-                          .join(''),
-              );
+            ? lapisClient.getSequenceFasta(accessionVersion)
+            : lapisClient.getMultiSegmentSequenceFasta(accessionVersion, segmentNames);
     },
 );
