+++ website/src/services/lapisClient.ts
@@ -21,6 +21,7 @@ import {
     type SequenceEntryHistory,
     versionStatuses,
 } from '../types/lapis.ts';
+import { fastaEntryToString, parseFasta } from '../utils/parseFasta.ts';
 import type { BaseType } from '../utils/sequenceTypeHelpers.ts';
 
 export class LapisClient extends ZodiosWrapperClient<typeof lapisApi> {
@@ -53,7 +54,7 @@ export class LapisClient extends ZodiosWrapperClient<typeof lapisApi> {
         });
     }
 
-    public getSequenceEntryVersionDetailsTsv(accessionVersion: string) {
+    public getSequenceEntryVersionDetailsTsv(accessionVersion: string): Promise<Result<string, ProblemDetail>> {
         return this.call('details', {
             [this.schema.primaryKey]: accessionVersion,
             dataFormat: 'TSV',
@@ -164,5 +165,31 @@ export class LapisClient extends ZodiosWrapperClient<typeof lapisApi> {
             ),
         );
         return Result.combine(results);
+    }
+
+    public getSequenceFasta(accessionVersion: string): Promise<Result<string, ProblemDetail>> {
+        return this.getUnalignedSequences(accessionVersion);
+    }
+
+    public async getMultiSegmentSequenceFasta(
+        accessionVersion: string,
+        segmentNames: string[],
+    ): Promise<Result<string, ProblemDetail>> {
+        const segments = await this.getUnalignedSequencesMultiSegment(accessionVersion, segmentNames);
+        return segments.map((segmentFastas) =>
+            segmentFastas
+                .map((fasta, i) => {
+                    const parsed = parseFasta(fasta);
+                    if (parsed.length === 0) {
+                        return '';
+                    }
+                    const withSegmentSuffix = {
+                        name: `${parsed[0].name}_${segmentNames[i]}`,
+                        sequence: parsed[0].sequence,
+                    };
+                    return fastaEntryToString([withSegmentSuffix]);
+                })
+                .join(''),
+        );
     }
 }
